backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
aiofiles==24.1.0
openai-whisper==20240930
pydantic==2.10.4
